improved setup mode detection

+++ app/app/routes/application.js
@@ -31,7 +31,9 @@ export default Ember.Route.extend(ApplicationRouteMixin, TooltipMixin, {
 
 		return this.get('appMeta').boot(transition.targetName, window.location.href).then(data => {
 			if (sa !== "authenticator:documize" && sa !== "authenticator:keycloak" && data.allowAnonymousAccess) {
-				return this.get('session').authenticate('authenticator:anonymous', data);
+				if (!this.get('appMeta.setupMode')) {
+					return this.get('session').authenticate('authenticator:anonymous', data);
+				}
 			}
 
 			return;
@@ -39,7 +41,12 @@ export default Ember.Route.extend(ApplicationRouteMixin, TooltipMixin, {
 	},
 
 	sessionAuthenticated() {
+		if (this.get('appMeta.setupMode')) {
+			return;
+		}
+
 		let next = this.get('localStorage').getSessionItem('entryUrl');
+
 		if (is.not.null(next) && is.not.undefined(next)) {
 			this.get('localStorage').clearSessionItem('entryUrl')
 
@@ -60,7 +67,7 @@ export default Ember.Route.extend(ApplicationRouteMixin, TooltipMixin, {
 				console.log(error); // eslint-disable-line no-console
 				console.log(transition); // eslint-disable-line no-console
 
-				if (netUtil.isAjaxAccessError(error)) {
+				if (netUtil.isAjaxAccessError(error) && !this.get('appMeta.setupMode')) {
 					localStorage.clear();
 					return this.transitionTo('auth.login');
 				}
